+++ WordCloudGenerator/WordFrequency.cs
@@ -2,13 +2,13 @@
 {
     public readonly struct WordFrequency
     {
-        public WordFrequency(string word, double frequency)
+        public WordFrequency(string word, float frequency)
         {
             Word = word;
             Frequency = frequency;
         }
 
         public string Word { get; }
-        public double Frequency { get; }
+        public float Frequency { get; }
     }
 }
